Preliminary refactoring, eliminated duplicate files with minor differences, applied OOP to seperate function into own file.
